Factor out data-generating utilities from file that writes to db
